LPS-57199 Move tag classes to the correct package
