--- modules/apps/staging/staging-taglib/src/com/liferay/staging/taglib/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
