+++ modules/apps/staging/staging-taglib/src/com/liferay/staging/taglib/servlet/taglib/ConfigurationHeaderTag.java
@@ -12,9 +12,10 @@
  * details.
  */
 
-package com.liferay.staging.taglib;
+package com.liferay.staging.taglib.servlet.taglib;
 
-import com.liferay.staging.taglib.util.ServletContextUtil;
+import com.liferay.portlet.exportimport.model.ExportImportConfiguration;
+import com.liferay.staging.taglib.servlet.ServletContextUtil;
 import com.liferay.taglib.util.IncludeTag;
 
 import javax.servlet.http.HttpServletRequest;
@@ -23,14 +24,16 @@ import javax.servlet.jsp.PageContext;
 /**
  * @author Levente Hudák
  */
-public class DeletionsTag extends IncludeTag {
+public class ConfigurationHeaderTag extends IncludeTag {
 
-	public void setCmd(String cmd) {
-		_cmd = cmd;
+	public void setExportImportConfiguration(
+		ExportImportConfiguration exportImportConfiguration) {
+
+		_exportImportConfiguration = exportImportConfiguration;
 	}
 
-	public void setDisableInputs(boolean disableInputs) {
-		_disableInputs = disableInputs;
+	public void setLabel(String label) {
+		_label = label;
 	}
 
 	@Override
@@ -42,8 +45,8 @@ public class DeletionsTag extends IncludeTag {
 
 	@Override
 	protected void cleanUp() {
-		_cmd = null;
-		_disableInputs = false;
+		_exportImportConfiguration = null;
+		_label = null;
 	}
 
 	@Override
@@ -53,14 +56,19 @@ public class DeletionsTag extends IncludeTag {
 
 	@Override
 	protected void setAttributes(HttpServletRequest request) {
-		request.setAttribute("liferay-staging:deletions:cmd", _cmd);
 		request.setAttribute(
-			"liferay-staging:deletions:disableInputs", _disableInputs);
+			"liferay-staging:configuration-header:exportImportConfiguration",
+			_exportImportConfiguration);
+		request.setAttribute(
+			"liferay-staging:configuration-header:label", _label);
+		request.setAttribute(
+			"liferay-staging:configuration-header:liferayPortletRequest",
+			pageContext.getAttribute("liferayPortletRequest"));
 	}
 
-	private static final String _PAGE = "/taglib/deletions/page.jsp";
+	private static final String _PAGE = "/taglib/configuration_header/page.jsp";
 
-	private String _cmd;
-	private boolean _disableInputs;
+	private ExportImportConfiguration _exportImportConfiguration;
+	private String _label;
 
 }
